fix: Fix type of predictions.

+++ models/detector.py
@@ -214,7 +214,7 @@ class PromptableDeTRTrainer(PromptableDeTR):
         logger.debug(msg="- New scores shape: %s." % (new_scores.shape,))
         new_scores[(batch_idx, tgt_idx)] = sorted_true_presence
 
-        new_pred_scores = torch.full(size=pred_presence.shape, fill_value=0, device=sorted_pred_presence.device).long()
+        new_pred_scores = torch.full(size=pred_presence.shape, fill_value=0.0, device=sorted_pred_presence.device).float()
         logger.debug(msg="- New predicted scores shape: %s." % (new_pred_scores.shape,))
         new_pred_scores[(batch_idx, src_idx)] = sorted_pred_presence
